More Fun!
The game is now more fun! Use AWSD to move.  Collect the capsule for a
power up!  Just playing around.  Added big textures I know.  Feel free
to play around with things.  Once we are ready to start building
something we can create a new scene and be more structured.